PTB BETA RELEASE "Hammertime!" SP3 - "Flowerpower"

Also removed manually
  ./Psychtoolbox/PsychDocumentation/PTBTutorial-ECVP2013.pdf
since no original source avail ATM

Fixes Linux X11 display locking bugs which could cause deadlock in
Screen('Resolutions'), and crashes for Screen('AsyncFlipBegin') async
flips and framesequential stereo mode 11. These bugs could only affect
the Mesa/Gallium FOSS graphics drivers, therefore slipped through the
testing which was performed with the NVidia binary blob. Retesting on
nouveau exposed the bugs.

-> Our locking bugs should be fixed without degradation in performance.
-> There are other bugs which can cause crashes at least with
   MultiWindowLockStepTest(n) for n > 1. These look like bugs in nouveau,
-> This release was validated on the upcoming Ubuntu 13.10 "Saucy Salamander"
   64-Bit Beta with nouveau.

* tag 'PTB_Beta-2013-10-3_V3.0.11': (141 commits)
  Rebuild Screen mex files for 32/64-Bit Octave and Matlab on Linux.
  Screen/Linux: Always poll-wait in PsychOSGetSwapCompletionTimestamp()
  Screen/Linux: Next attempt at proper X11 Locking :-(
  Screen/Linux: Fix deadlock in Screen('Resolutions', screenid);
  Screen/Linux: Fix X11 display locking more.
  GazeContingentDemo - fix small bug.
  BitsPlusImagingPipelineTest: Make more lenient for acceptable errors.
  SuperShapeDemo: Disable on ARM OpenGL-ES due to lack of display lists.
  Screen, PsychHID, ftgles drawtext plugin rebuilt for Nexus/ARM Linux.
  help Psychtoolbox reformatted and clarified.
  Download/UpdatePsychtoolobx - Update and reformat help text.
  help BeampositionQueries and help SyncTrouble - update and reformat.
  Rebuild all Screen mex files for multisampling + display rotation support.
  Screen: Fix multisample texture support for good.
  Screen/OSX: Fix Intel + No beampos for good.
  PanelFitterDemo: Add test code for multisampling + rotation.
  Screen mex file for all desktop platforms rebuilt.
  Rebuild ftgl drawtext text rendering plugins for desktop Linux and OSX.
  Screen: Disable beamposition queries for Intel gpu's under OSX and Linux.
  Screen: Allow use of multisample textures for multisample anti-aliasing.
  ...

+++ Psychtoolbox/PsychGLImageProcessing/PsychVideoSwitcher.m
@@ -20,8 +20,11 @@ function varargout = PsychVideoSwitcher(cmd, varargin)
 % standard RGB true color display mode. 'screenIdx' is the screen index of the
 % for the display screen to switch.
 %
-% 'enableLuminanceMode' must be set to 0 to switch to RGB mode, and to 1 to
-% switch to high precision luminance mode.
+% 'enableLuminanceMode' is meaningful only for card VideoSwitcher, which
+% must be set to 0 to switch to RGB mode, and to 1 to switch to high
+% precision luminance mode. For box version, the SwitchMode subfunction
+% ignores the third input, and only toggles between two display modes,
+% equivalent to pushing the switch button on the box.
 %
 % 'VideoSwitcherIsABox' is an optional argument: If set to 1, then perform
 % switching procedure for an external (box) device. If set to zero, then
@@ -181,6 +184,8 @@ function varargout = PsychVideoSwitcher(cmd, varargin)
 %              Xiangrui Li in helper subroutines.
 % 05/24/11 xl  3rd input for SwitchMode is not manditory since it is not
 %              used by box version.
+% 07/17/13 xl  Clarify in help text that SwitchMode only toggles between
+%              two display modes.  
 
 % GL access is needed for setup of green trigger channel in callback:
 global GL;
@@ -787,8 +792,8 @@ for i=1:25
     img1(1, pixelPos(i)+1:pixelPos(i+1), 2)=bits(1,i)*255;
     img2(1, pixelPos(i)+1:pixelPos(i+1), 2)=bits(2,i)*255;
 end
-tex(1)=Screen('MakeTexture',w,img1);
-tex(2)=Screen('MakeTexture',w,img2);
+tex(1)=Screen('MakeTexture',w,uint8(img1));
+tex(2)=Screen('MakeTexture',w,uint8(img2));
 
 for i=1:2
     for factor=(range(1):step:range(2))/range(1)
@@ -838,7 +843,7 @@ img=zeros(1, pixelPos(26), 3); % one line image, will change green layer
 for i=1:25
     img(1, pixelPos(i)+1:pixelPos(i+1), 2)=bits(color,i)*255;
 end
-tex=Screen('MakeTexture',w,img);
+tex=Screen('MakeTexture',w,uint8(img));
 
 for factor=(range(1):step:range(2))/range(1)
     rd=rand*ppb*0.5; % random shift up to half a bit
